// src/nlpviewer/src/lib/interfaces.ts
export interface IAttributes {
  [key: string]: any;
}

export interface ISpan {
  begin: number;
  end: number;
}

export interface IEntry {
  id: string;
  attributes: IAttributes;
}

export interface IAnnotation extends IEntry {
  legendId: string;
  span: ISpan;
}
// ...
export interface IPlugin {
  name: string;
  component: (props: any) => JSX.Element | null;
  enabled: (props: any) => boolean;
}

export interface ILayout {
  [position: string]: string;
}

export interface ILegendAttributeConfig {
  [attribute: string]: boolean;
}

export interface ILegendConfig {
  is_selected: boolean;
  is_shown: boolean;
  attributes?: ILegendAttributeConfig;
}

export interface ILegendConfigs {
  [key: string]: ILegendConfig;
}

export interface IScopeConfigs {
  [key: string]: boolean;
}

export interface IProjectConfigs {
  legendConfigs: ILegendConfigs;
  scopeConfigs: IScopeConfigs;
  layoutConfigs: ILayout;
}

export interface IDocument {
  id: number;
  name?: string;
}

export interface IDocuments {
  documents: IDocument[];
}